+++ receiver/flinkmetricsreceiver/internal/metadata/generated_metrics_v2.go
@@ -547,7 +547,7 @@ type metricFlinkJvmCPULoad struct {
 // init fills flink.jvm.cpu.load metric with initial data.
 func (m *metricFlinkJvmCPULoad) init() {
 	m.data.SetName("flink.jvm.cpu.load")
-	m.data.SetDescription("The CPU usage (%) of the JVM for a jobmanager or taskmanager.")
+	m.data.SetDescription("The CPU usage of the JVM for a jobmanager or taskmanager.")
 	m.data.SetUnit("1")
 	m.data.SetDataType(pmetric.MetricDataTypeGauge)
 }